+++ test/tests.test.js
@@ -5,9 +5,9 @@ describe('nobind', function () {
   for (const t of tests) {
     describe(t, () => {
       framework.register(t);
-      before('configure', () => framework.configure(t));
+      before('configure', () => framework.configure(t, undefined, process.env.ENABLE_ASAN && ['--debug', '--enable_asan']));
       before('build', () => framework.build());
-      before('load', () => framework.load(t));
+      before('load', () => framework.load(t, process.env.ENABLE_ASAN && 'Debug'));
       after('GC', global.gc);
     });
   }