+++ include/nobind.h
@@ -67,15 +67,28 @@ public:
 
 } // namespace Nobind
 
-#define NOBIND_MODULE(MODULE_NAME, MODULE_ARG)                                                                         \
+#define _NOBIND_MODULE_3(MODULE_NAME, MODULE_ARG, INSTANCE_DATA_TYPE)                                                  \
+  /* We make this assumption that is not guaranteed by the C++ specs but it seems to be true on all compilers */       \
+  static_assert(&Nobind::BaseEnvInstanceData::cons == &Nobind::EnvInstanceData<INSTANCE_DATA_TYPE>::cons);             \
   char const Nobind_##MODULE##_name[] = #MODULE_NAME;                                                                  \
   void Nobind_##MODULE##_Init_Wrapper(Nobind::Module<Nobind_##MODULE##_name> &);                                       \
   Napi::Object Nobind_##MODULE##_Init_Wrapper(Napi::Env, Napi::Object);                                                \
   NODE_API_MODULE(MODULE_NAME, Nobind_##MODULE##_Init_Wrapper)                                                         \
   Napi::Object Nobind_##MODULE##_Init_Wrapper(Napi::Env env, Napi::Object exports) {                                   \
-    env.SetInstanceData(new Nobind::EnvInstanceData);                                                                  \
+    env.SetInstanceData(new Nobind::EnvInstanceData<INSTANCE_DATA_TYPE>);                                              \
     Nobind::Module<Nobind_##MODULE##_name> m(env, exports);                                                            \
     Nobind_##MODULE##_Init_Wrapper(m);                                                                                 \
     return exports;                                                                                                    \
   }                                                                                                                    \
   void Nobind_##MODULE##_Init_Wrapper(Nobind::Module<Nobind_##MODULE##_name> &MODULE_ARG)
+
+#define _NOBIND_MODULE_2(MODULE_NAME, MODULE_ARG)                                                                      \
+  _NOBIND_MODULE_3(MODULE_NAME, MODULE_ARG, Nobind::EmptyEnvInstanceData)
+#define _NOBIND_MODULE_1(MODULE_NAME) _NOBIND_MODULE_2(MODULE_NAME, m)
+
+// macro with optional arguments
+// from https://stackoverflow.com/questions/3046889/optional-parameters-with-c-macros
+#define _NOBIND_GET_4TH_ARG(arg1, arg2, arg3, arg4, ...) arg4
+#define _NOBIND_MODULE_MACRO_CHOOSER(...)                                                                              \
+  _NOBIND_GET_4TH_ARG(__VA_ARGS__, _NOBIND_MODULE_3, _NOBIND_MODULE_2, _NOBIND_MODULE_1, )
+#define NOBIND_MODULE(...) _NOBIND_MODULE_MACRO_CHOOSER(__VA_ARGS__)(__VA_ARGS__)