+++ CHANGELOG.md
@@ -5,9 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### []
+
+-   Specify the memory ownership rules for `Buffer`
+-   Fix a minor memory leak when calling method with incorrect number of arguments
+
 ### [1.1.1] 2023-12-01
 
- -  Fix [#7](https://github.com/mmomtchev/nobind/issues/7), escape the include directory path on Windows
+-   Fix [#7](https://github.com/mmomtchev/nobind/issues/7), escape the include directory path on Windows
 
 ## [1.1.0] 2023-11-30